refactor(post): 공개 조회 응답 정책 컴포넌트 분리
